requirements.txt
asgiref==3.5.2
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
distlib==0.3.4
Django==3.2.13
django-rest-framework==0.1.0
djangorestframework==3.13.1
dnspython==2.2.1
email-validator==1.2.1
filelock==3.7.1
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
nodejs==0.1.1
npm==0.1.1
optional-django==0.1.0
platformdirs==2.5.2
pycparser==2.21
pytz==2022.1
requests==2.28.0
six==1.16.0
SQLAlchemy==1.4.35
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
Pillow==9.1.1